backend/pyproject.toml
[project]
name = "deer-flow"
version = "0.1.0"
description = "LangGraph-based AI agent system with sandbox execution capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-sandbox>=0.0.19",
    "dotenv>=0.9.9",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "kubernetes>=30.0.0",
    "langchain>=1.2.3",
    "langchain-anthropic>=0.3.0",
    "langchain-deepseek>=1.0.1",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "langgraph-cli[inmem]>=0.4.11",
    "markdownify>=1.2.2",
    "markitdown[all,xlsx]>=0.0.1a2",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "readabilipy>=0.3.0",
    "sse-starlette>=2.1.0",
    "tavily-python>=0.7.17",
    "firecrawl-py>=1.15.0",
    "tiktoken>=0.8.0",
    "uvicorn[standard]>=0.34.0",
    "ddgs>=9.10.0",
    "duckdb>=1.4.4",
    "bcrypt>=4.2.0",
    "PyJWT>=2.10.0",
    "pydantic[email]>=2.12.5",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "gunicorn>=22.0.0",
    "redis>=5.0.0",
    "rq>=1.16.0",
    "python-json-logger>=2.0.0",
    "prometheus-client>=0.20.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
]

[dependency-groups]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0", "httpx>=0.28.0", "ruff>=0.14.11"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require Docker or external services (deselect with '-m \"not integration\"')",
]
